pyproject.toml
[project]
name = "qsb-serial"
version = "0.1.0"
description = "Serial communication library for QSB devices"
authors = [
    { name = "Gergely Turi", email = "gt2253@cumc.columbia.edu" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
]
